+++ NEWS.md
@@ -2,10 +2,15 @@
 title: "NEWS"
 output: html_document
 ---
+## backbone 3.0.4
+
+* perform parameter/input checks inside each `backbone_from_` function
+* updated citation
+
 ## backbone 3.0.3
 
-* Updated/added generic `print()`, `summary()` and `plot()` functions for backbone objects
-* Corrected p-values in asymmetric marginal likelihood filter (h/t @florian-huber)
+* updated/added generic `print()`, `summary()` and `plot()` functions for backbone objects
+* corrected p-values in asymmetric marginal likelihood filter (h/t @florian-huber)
 
 ## backbone 3.0.2